Turning Angles + Speed Input
Tuned the turning and forward angles so that it travels in the expected directions and doesn't encounter issues with the turning angle being too sharp and interfering with movement. Additionally, added a prompt when the 'S' key is pressed that allows the user to change the speed of the robot between 0-100.

Could not figure out how to clear the input field when asking for the speed, so the last answer (plus the 'S' that prompted the input) are recorded in the input field and must be cleared before entering the desired speed (is there a way to fix this?). We also tried to smooth out some of the car motion with questionable levels of success, so that it would behave more like a remote control car.