Improved Controlling + Speed Validation
Attached is an improved client side file that reworks the controlling scheme so that the car controls like a remote control car (a more intuitive control model than what we had before). It includes the on_release_key so that when a key is released, the command stops. Additionally, steering is made independent of the forward/backward motion allowing for more smooth driving. The speed selection prompt was also updated to sift out input that is beyond the valid range or valid input type.

There is a corresponding picar_server_improved file on the picar that runs in conjunction with the attached file to make the above changes possible.

Note - The 'B' key was removed from the list of usable keys in this improved file as one only needs to release the forward/backward arrow to stop.